refactor: move into schema
