+++ magicgui/_version.py
@@ -1,0 +1,4 @@
+# file generated by setuptools_scm
+# don't change, don't track in version control
+__version__ = version = "0.6.1.dev13+g14d82dc"
+__version_tuple__ = version_tuple = (0, 6, 1, "dev13", "g14d82dc")
