+++ magicgui/schema/_guiclass.py
